Move regrid_method_type to common utilities
