+++ imod/mf6/dis.py
@@ -12,10 +12,11 @@ from imod.common.utilities.regrid import (
     _regrid_like,
     _regrid_package_data,
 )
+from imod.common.utilities.regrid_method_type import RegridMethodType
 from imod.logging import init_log_decorator, standard_log_decorator
 from imod.mf6.disv import VerticesDiscretization
 from imod.mf6.package import Package
-from imod.mf6.regrid.regrid_schemes import DiscretizationRegridMethod, RegridMethodType
+from imod.mf6.regrid.regrid_schemes import DiscretizationRegridMethod
 from imod.mf6.utilities.grid import create_smallest_target_grid
 from imod.mf6.utilities.imod5_converter import convert_ibound_to_idomain
 from imod.mf6.validation import DisBottomSchema
