+++ imod/msw/initial_conditions.py
@@ -3,9 +3,9 @@ import shutil
 from typing import Any, TextIO
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
+from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
-from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
 
 
 class InitialConditionsEquilibrium(MetaSwapPackage, IRegridPackage):
