+++ imod/msw/pkgbase.py
@@ -11,12 +11,12 @@ from imod.common.utilities.regrid import (
     RegridderWeightsCache,
     _regrid_like,
 )
+from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
 from imod.mf6.dis import StructuredDiscretization
 from imod.mf6.mf6_wel_adapter import Mf6Wel
 from imod.msw.fixed_format import format_fixed_width
 from imod.typing import IntArray
 from imod.typing.grid import GridDataArray, GridDataset
-from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
 
 DataDictType: TypeAlias = dict[str, IntArray | int | str]
