+++ imod/msw/vegetation.py
@@ -4,9 +4,9 @@ import numpy as np
 import xarray as xr
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
+from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
-from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
 
 
 class AnnualCropFactors(MetaSwapPackage, IRegridPackage):
