+++ imod/util/regrid_method_type.py
@@ -1,11 +1,6 @@
 from enum import Enum
-from typing import ClassVar, Protocol, Tuple, TypeAlias
 
 import xugrid as xu
-from pydantic import ConfigDict
-from pydantic.dataclasses import dataclass
-
-_CONFIG = ConfigDict(extra="forbid")
 
 
 class RegridderType(Enum):
@@ -20,26 +15,3 @@ class RegridderType(Enum):
     BARYCENTRIC = xu.BarycentricInterpolator
     OVERLAP = xu.OverlapRegridder
     RELATIVEOVERLAP = xu.RelativeOverlapRegridder
-
-
-_RegridVarType: TypeAlias = Tuple[RegridderType, str] | Tuple[RegridderType]
-
-
-class RegridMethodType(Protocol):
-    # Work around that type annotation is a bit hard on dataclasses, as they
-    # don't expose a class interface.
-    # Adapted from: https://stackoverflow.com/a/55240861
-    # "As already noted in comments, checking for this attribute is currently the
-    # most reliable way to ascertain that something is a dataclass"
-    # See also:
-    # https://github.com/python/mypy/issues/6568#issuecomment-1324196557
-
-    __dataclass_fields__: ClassVar[dict]
-
-    def asdict(self) -> dict:
-        return vars(self)
-
-
-@dataclass(config=_CONFIG)
-class EmptyRegridMethod(RegridMethodType):
-    pass
